refactor: move TaskType to constants
